+++ src/components/ChannelPreview/hooks/useLatestMessagePreview.ts
@@ -7,7 +7,6 @@ import {
   useTranslationContext,
 } from '../../../contexts/translationContext/TranslationContext';
 
-import type { Immutable } from 'seamless-immutable';
 import type {
   Channel,
   ChannelState,
@@ -35,10 +34,7 @@ type LatestMessage<
   Re extends UnknownType = DefaultReactionType,
   Us extends UnknownType = DefaultUserType
 > =
-  | Immutable<
-      ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['messageToImmutable']>
-    >
-  | ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['messageToImmutable']>
+  | ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['formatMessage']>
   | MessageResponse<At, Ch, Co, Me, Re, Us>;
 
 export type LatestMessagePreview<
@@ -150,13 +146,7 @@ const getLatestMessageDisplayDate = <
   message: LatestMessage<At, Ch, Co, Ev, Me, Re, Us> | undefined,
   tDateTimeParser: TDateTimeParser,
 ) => {
-  const parserOutput = tDateTimeParser(
-    message?.created_at
-      ? typeof message.created_at === 'string'
-        ? message.created_at
-        : message.created_at.asMutable()
-      : undefined,
-  );
+  const parserOutput = tDateTimeParser(message?.created_at);
   if (isDayOrMoment(parserOutput)) {
     if (parserOutput.isSame(new Date(), 'day')) {
       return parserOutput.format('LT');
@@ -190,7 +180,7 @@ const getLatestMessageReadStatus = <
   if (!message || currentUserId !== message.user?.id || readEvents === false)
     return 0;
 
-  const readList = channel.state.read.asMutable();
+  const readList = channel.state.read;
   if (currentUserId) {
     delete readList[currentUserId];
   }
@@ -223,7 +213,7 @@ const getLatestMessagePreview = <
   t: (key: string) => string;
   tDateTimeParser: TDateTimeParser;
   lastMessage?:
-    | ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['messageToImmutable']>
+    | ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['formatMessage']>
     | MessageResponse<At, Ch, Co, Me, Re, Us>;
 }) => {
   const {
@@ -280,7 +270,7 @@ export const useLatestMessagePreview = <
   channel: Channel<At, Ch, Co, Ev, Me, Re, Us>,
   forceUpdate: number,
   lastMessage?:
-    | ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['messageToImmutable']>
+    | ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['formatMessage']>
     | MessageResponse<At, Ch, Co, Me, Re, Us>,
 ) => {
   const { client } = useChatContext<At, Ch, Co, Ev, Me, Re, Us>();
@@ -291,7 +281,9 @@ export const useLatestMessagePreview = <
   const messages = channel.state.messages;
   const message = messages.length ? messages[messages.length - 1] : undefined;
 
-  const lastMessageId = lastMessage?.id || message?.id;
+  const channelLastMessageString = `${lastMessage?.id || message?.id}${
+    lastMessage?.updated_at || message?.updated_at
+  }`;
 
   const [readEvents, setReadEvents] = useState(true);
   const [latestMessagePreview, setLatestMessagePreview] = useState<
@@ -336,7 +328,7 @@ export const useLatestMessagePreview = <
           tDateTimeParser,
         }),
       ),
-    [forceUpdate, lastMessageId, readEvents, readStatus],
+    [channelLastMessageString, forceUpdate, readEvents, readStatus],
   );
 
   return latestMessagePreview;