+++ src/components/MessageInput/MessageInput.tsx
@@ -364,9 +364,11 @@ export const MessageInputWithContext = <
   const getMembers = () => {
     const result: UserResponse<Us>[] = [];
     if (members && Object.values(members).length) {
-      Object.values(members).forEach((member) =>
-        result.push(member.user as UserResponse<Us>),
-      );
+      Object.values(members).forEach((member) => {
+        if (member.user) {
+          result.push(member.user);
+        }
+      });
     }
 
     return result;
@@ -379,7 +381,7 @@ export const MessageInputWithContext = <
     const uniqueUsers: { [key: string]: UserResponse<Us> } = {};
     for (const user of users) {
       if (user && !uniqueUsers[user.id]) {
-        uniqueUsers[user.id] = user as UserResponse<Us>;
+        uniqueUsers[user.id] = user;
       }
     }
     const usersArray = Object.values(uniqueUsers);
@@ -390,7 +392,7 @@ export const MessageInputWithContext = <
   const getWatchers = () => {
     const result: UserResponse<Us>[] = [];
     if (watchers && Object.values(watchers).length) {
-      result.push(...(Object.values(watchers) as UserResponse<Us>[]));
+      result.push(...Object.values(watchers));
     }
 
     return result;